Move modules to src package and update imports
